Implement Scratch password resets
All over:
* Make all methods just functions, instead of static/public/private
* loginForm renamed to showForm as it's not just login anymore

+ScratchLogin.common.php:
* Move helper functions and definitions to this file
+ ScratchSpecialPage centralizes form-showing, verifying, code-resetting logic
  * Defines execute which calls showForm, onPost, resetCode
  * Defines verifForm which showForm implementations call
  * Defines verifSucceeded which onPost implementations call
    * Returns a User object or null
  * Defines doCodeReset which resetCode implementations call

SpecialScratchLogin.php:
- Move helper functions and definitions to ScratchLogin.common.php
* class SpecialScratchLogin extends ScratchSpecialPage
- Form showing is now in ScratchSpecialPage
- Verification is now in ScratchSpecialpage
- Code reset   is now in ScratchSpecialPage
- Move hook into ScratchLogin.hooks.php

+SpecialScratchPasswordReset.php:
* Form is mostly the same as ScratchLogin
* The five functional lines of code in onPost took me six hours of manic
  reading of MediaWiki class documentation and source code which occupied
  at least fifty different tabs
* Resets password and requires changing it after logging in with it

+ScratchLogin.hooks.php:
* Move hook into this file
+ Link to Special:ScratchPasswordReset from Special:PasswordReset

extension.json:
+ Autoload SpecialScratchPasswordReset
+ Autoload ScratchLoginHooks
+ Special:ScratchPasswordReset is defined by SpecialScratchPasswordReset
* Hooks now come from ScratchLoginHooks

i18n/en.json:
* scratchlogin-instructions: oops, forgot to close </small>
* scratchlogin-code-reset: This now has a parameter for the return link,
  which is the only thing resetCode in ScratchSpecialPages provides
+ scratchpasswordreset: title of special page
+ scratchpasswordreset-instructions: mostly the same as login
+ scratchpasswordreset-verify: Text of the button since "Log in" doesn't work
+ scratchpasswordreset-success: Provides the temp pw and links to log in
+ scratchpasswordreset-passwordreset-link: Link on original pw reset page

i18n/qqq.json:
* Documented en.json
