+++ i18n/en.json
@@ -5,16 +5,19 @@
 			"Kenny2scratch"
 		]
 	},
-	"scratchlogin": "Scratch Login",
-	"scratchlogin-desc": "Allows logging in with a Scratch account.",
-	"scratchlogin-title": "Log in with Scratch",
-	"scratchlogin-summary": "Log in with your Scratch account.",
+	"scratchlogin": "Log in with Scratch",
+	"scratchlogin-desc": "Allows logging in or resetting passwords with a Scratch account.",
 	"scratchlogin-project-id": "10135908",
 	"scratchlogin-project-author": "ModShare",
-	"scratchlogin-instructions": "'''Note: You must already have an account on this wiki to log in with Scratch!'''<br/>Please comment the following verification code on the $1user verification project$2:<br/>'''$3'''<br/><small>(If necessary, you can [[Special:ScratchLogin/reset|reset your verification code]].)",
+	"scratchlogin-instructions": "'''Note: You must already have an account on this wiki to log in with Scratch!'''<br/>Please comment the following verification code on the $1user verification project$2:<br/>'''$3'''<br/><small>(If necessary, you can [[Special:ScratchLogin/reset|reset your verification code]].)</small>",
 	"scratchlogin-uncommented": "You do not appear to have commented the code on the project.",
 	"scratchlogin-unregistered": "The user '''$1''' is not registered on this wiki.",
 	"scratchlogin-success": "Success! You will be logged in as '''$1''' on the next page you visit.",
-	"scratchlogin-code-reset": "Your verification code has been reset.<br/>Return to [[Special:ScratchLogin]].",
-	"scratchlogin-userlogin-link": "<p class=\"mw-ui-vform\" style=\"text-align: center\">[[Special:ScratchLogin|Log in with Scratch]]</p>"
+	"scratchlogin-code-reset": "Your verification code has been reset.<br/>Return to [[$1]].",
+	"scratchlogin-userlogin-link": "<p class=\"mw-ui-vform\" style=\"text-align: center;\">[[Special:ScratchLogin|Log in with Scratch]]</p>",
+	"scratchpasswordreset": "Reset password with Scratch",
+	"scratchpasswordreset-instructions": "Comment the following verification code on the $1user verification project$2 to be displayed a temporary password to log in with:<br/>'''$3'''<br/><small>(If necessary, you can [[Special:ScratchPasswordReset/reset|reset your verification code]].)</small>",
+	"scratchpasswordreset-verify": "Verify",
+	"scratchpasswordreset-success": "Your temporary password is '''<code>$1</code>'''. Please [[Special:UserLogin/$2|log in]] and change it as soon as possible.",
+	"scratchpasswordreset-passwordreset-link": "[[Special:ScratchPasswordReset|Reset your password with Scratch]]"
 }
