+++ i18n/qqq.json
@@ -15,6 +15,11 @@
 	"scratchlogin-uncommented": "Error message when the code is not found.",
 	"scratchlogin-unregistered": "Error message when the user who commented the code is not registered on the wiki. $1 is the username found.",
 	"scratchlogin-success": "Message for successful login. $1 is the logged in username.",
-	"scratchlogin-code-reset": "Message for successful reset of verification code.",
-	"scratchlogin-userlogin-link": "Link added to [[Special:UserLogin]] linking to [[Special:ScratchLogin]]."
+	"scratchlogin-code-reset": "Message for successful reset of verification code. $1 is the special page to return to.",
+	"scratchlogin-userlogin-link": "Link added to [[Special:UserLogin]] linking to [[Special:ScratchLogin]].",
+	"scratchpasswordreset": "Title of Special:ScratchPasswordReset",
+	"scratchpasswordreset-instructions": "Instructions shown on [[Special:ScratchPasswordReset]]. Parameters are the same as scratchlogin-instructions.",
+	"scratchpasswordreset-verify": "Text of \"Verify\" button used instead of \"Log in\".",
+	"scratchpasswordreset-success": "Message for successful password reset. $1 is the temporary password and $2 is the username of the user whose password was reset, used to link to UserLogin with that username.",
+	"scratchpasswordreset-passwordreset-link": "Link added to [[Special:PasswordReset]] linking to [[Special:ScratchPasswordReset]]."
 }
